+++ examples/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @docknetwork/sdk-examples
 
+## 0.10.0
+
+### Minor Changes
+
+- Legacy documents support for `testnet`
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/dock-blockchain-modules@0.16.0
+  - @docknetwork/dock-blockchain-api@0.12.0
+  - @docknetwork/credential-sdk@0.27.0
+
 ## 0.9.0
 
 ### Minor Changes