+++ packages/credential-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/credential-sdk
 
+## 0.27.0
+
+### Minor Changes
+
+- Legacy documents support for `testnet`
+
 ## 0.26.0
 
 ### Minor Changes