+++ packages/dock-blockchain-api/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-api
 
+## 0.12.0
+
+### Minor Changes
+
+- Legacy documents support for `testnet`
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.27.0
+
 ## 0.11.0
 
 ### Minor Changes